Revert "Expose Snap Threshold and Move Snap Settings to BBox Panel"
We will handle this a bit differently - by allowing the grid origin to be moved. I will dig in at some point.

This reverts commit 33c92ecf4da724c2f17d9d91c7ea31a43a2f6deb.
